+++ ld/document_loader_test.go
@@ -16,7 +16,7 @@ package ld_test
 
 import (
 	"bytes"
-	. "github.com/piprate/json-gold/ld"
+	. "github.com/soluchok/json-gold/ld"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"testing"